+++ lib/open_graph_reader/object/dsl/types.rb
@@ -19,9 +19,11 @@ module OpenGraphReader
       define_type_no_doc :url do |value, options|
         value = value.to_s
 
+
         unless value.start_with?('http://') || value.start_with?('https://')
           if options[:image] && OpenGraphReader.config.synthesize_image_url
             unless OpenGraphReader.current_origin
+              next unless options[:required] || !OpenGraphReader.config.discard_invalid_optional_properties
               raise ArgumentError, "Enabled image url synthesization but didn't pass an origin"
             end
 
@@ -35,10 +37,12 @@ module OpenGraphReader
                 uri.path = value
                 value = uri.to_s
               rescue
+                next unless options[:required] || !OpenGraphReader.config.discard_invalid_optional_properties
                 raise InvalidObjectError, "URL #{value.inspect} does not start with http:// or https:// and failed to synthesize a full URL"
               end
             end
           elsif options.has_key?(:to) && OpenGraphReader.config.validate_references
+            next unless options[:required] || !OpenGraphReader.config.discard_invalid_optional_properties
             raise InvalidObjectError, "URL #{value.inspect} does not start with http:// or https://"
           end
         end
@@ -50,46 +54,53 @@ module OpenGraphReader
       #   @param [Array<String>] allowed the list of allowed values
       #   @!macro define_type_description
       #   @see http://ogp.me/#enum
-      define_type_no_doc :enum do |value, allowed|
+      define_type_no_doc :enum do |value, allowed, options|
+        value = value.to_s
+
         unless allowed.include? value
+          next unless options[:required] || !OpenGraphReader.config.discard_invalid_optional_properties
           raise InvalidObjectError, "Expected one of #{allowed.inspect} but was #{value.inspect}"
         end
 
-        value.to_s
+        value
       end
 
       # @see http://ogp.me/#integer
-      define_type :integer do  |value|
+      define_type :integer do  |value, options|
         begin
           Integer(value)
         rescue  ArgumentError => e
+          next unless options[:required] || !OpenGraphReader.config.discard_invalid_optional_properties
           raise InvalidObjectError, "Integer expected, but was #{value.inspect}"
         end
       end
 
       # @see http://ogp.me/#datetime
-      define_type :datetime do |value|
+      define_type :datetime do |value, options|
         begin
           DateTime.iso8601 value
         rescue ArgumentError => e
+          next unless options[:required] || !OpenGraphReader.config.discard_invalid_optional_properties
           raise InvalidObjectError, "ISO8601 datetime expected, but was #{value.inspect}"
         end
       end
 
       # @see http://ogp.me/#bool
-      define_type :boolean do |value|
+      define_type :boolean do |value, options|
         {'true' => true, 'false' => false, '1' => true, '0' => false}[value].tap {|bool|
           if bool.nil?
+            next unless options[:required] || !OpenGraphReader.config.discard_invalid_optional_properties
             raise InvalidObjectError, "Boolean expected, but was #{value.inspect}"
           end
         }
       end
 
       # @see http://ogp.me/#float
-      define_type :float do |value|
+      define_type :float do |value, options|
         begin
           Float(value)
         rescue ArgumentError => e
+          next unless options[:required] || !OpenGraphReader.config.discard_invalid_optional_properties
           raise InvalidObjectError, "Float expected, but was #{value.inspect}"
         end
       end